Optimized day 12
Main Optimizations:

Switched from VeqDeqs and cloning to just operating on the slice
of the original Vec. "popping" just becomes a subslice operation.

Cleaned up the double recursive step in for the Unkown state.
This was made easier by splitting the Broken and Operational states
into their own functions.

The above made it so the cache keys could be represented by the
lengths of the slices instead of the slices themselves.

Because of the above, the cache key was now just 3 unsigned ints.
This allows use of an Array3 as the cache that is fully
preallocated and provides fast lookups with no hashing. Curiously,
this last optimization actually runs slower when running the
ConditionRecords in parallel, but is 3x faster in single threaded
than the parallel version with a Hashmap cache. (and 8x faster
than it's own parallel version).